Handle of non-partitioned Pulsar topics
